Mapstruct partial working

+++ assignment/src/main/java/com/mindtree/assignment/model/Product.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 public abstract class Product {
 
-	protected int productid;
+	protected long productid;
 	
 	protected String prodName;
